refactor(@n8n/expression-runtime): Split runtime/index.ts into focused modules
- safe-globals.ts: SafeObject, SafeError, ExpressionError, __sanitize
- lazy-proxy.ts: createDeepLazyProxy
- reset.ts: resetDataProxies
- index.ts: global declarations, library setup, wires everything to globalThis

esbuild bundles all modules regardless; the split is purely for readability.
